# server/openapi.yaml
openapi: 3.1.0
info:
  title: 采购看板 API
  version: 0.1.0
  description: |
    Procurement Kanban backend API.
    Use `/auth/login` (human) or hold a long-lived token issued by `/auth/agent-token` (machine / MCP).
    All authenticated requests must include `Authorization: Bearer <token>`.
servers:
  - url: http://localhost:3001/api
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /items:
    get:
      tags: [items]
      summary: List items (visibility-filtered)
      parameters:
        - in: query
          name: view
          schema: { type: string, enum: [personal, department] }
          description: Defaults to personal (only own items). Set to "department" to include team's DEPARTMENT-visible items.
        - in: query
          name: status
          schema: { type: string, enum: [TODO, IN_PROGRESS, COMPLETED] }
        - in: query
          name: priority
          schema: { type: string, enum: [URGENT, HIGH, MEDIUM, LOW] }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: projectId
          schema: { type: string }
          description: 'Pass the literal string "null" to filter inbox items (no project)'
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        '200':
          description: Items list (paginated object if page+limit, else flat array)
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: '#/components/schemas/Item' }
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Item' }
                      total: { type: integer }
                      page: { type: integer }
                      limit: { type: integer }
                      pages: { type: integer }
    post:
      tags: [items]
      summary: Create an item
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ItemCreate' }
      responses:
        '200':
          description: Created item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
  /items/summary:
    get:
      tags: [summary]
      summary: Aggregate financial + status summary (respects view + filters)
      parameters:
        - $ref: '#/components/parameters/View'
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Aggregated stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  totals: { type: object }
                  byCategory: { type: array, items: { type: object } }
                  byStatus: { type: array, items: { type: object } }
                  overBudgetItems: { type: array, items: { type: object } }
                  monthly: { type: object }
  /items/{id}:
    get:
      tags: [items]
      summary: Get a single item (includes statusHistories, reminders, project)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: Item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        '404': { description: Not found or not visible }
    put:
      tags: [items]
      summary: Update item (owner only)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ItemPatch' }
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
  /items/{id}/status:
    patch:
      tags: [items]
      summary: Update main status (writes StatusHistory automatically)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [TODO, IN_PROGRESS, COMPLETED] }
                subStatus: { type: string, nullable: true }
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
  /items/{id}/history:
    get:
      tags: [items]
      summary: Status change history (visibility-checked)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: History entries (desc by createdAt)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StatusHistory' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
    View:
      in: query
      name: view
      schema: { type: string, enum: [personal, department] }
    ProjectId:
      in: query
      name: projectId
      schema: { type: string }
  schemas:
...
    Item:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        progress: { type: string, nullable: true }
        status: { type: string, enum: [TODO, IN_PROGRESS, COMPLETED] }
        subStatus: { type: string, nullable: true }
        priority: { type: string, enum: [URGENT, HIGH, MEDIUM, LOW] }
        category: { type: string }
        dueDate: { type: string, format: date-time, nullable: true }
        visibility: { type: string, enum: [PRIVATE, DEPARTMENT, SHARED] }
        userId: { type: string }
        departmentId: { type: string, nullable: true }
        projectId: { type: string, nullable: true }
        estimatedAmount: { type: number, nullable: true }
        finalAmount: { type: number, nullable: true }
        currency: { type: string, nullable: true }
        supplierName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
    ItemCreate:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        priority: { type: string, enum: [URGENT, HIGH, MEDIUM, LOW] }
        category: { type: string }
        status: { type: string, enum: [TODO, IN_PROGRESS, COMPLETED] }
        subStatus: { type: string }
        dueDate: { type: string, format: date-time }
        visibility: { type: string, enum: [PRIVATE, DEPARTMENT, SHARED], default: DEPARTMENT }
        departmentId: { type: string }
        projectId: { type: string }
        estimatedAmount: { type: number }
        finalAmount: { type: number }
        currency: { type: string }
        supplierName: { type: string }
        supplierAmount: { type: number }
        requesterDepartment: { type: string }
    ItemPatch:
      allOf:
        - $ref: '#/components/schemas/ItemCreate'
      description: All fields are optional in PUT.
...
    StatusHistory:
      type: object
      properties:
        id: { type: string }
        itemId: { type: string }
        fromStatus: { type: string, nullable: true }
        toStatus: { type: string, nullable: true }
        fromSubStatus: { type: string, nullable: true }
        toSubStatus: { type: string, nullable: true }
        operatorId: { type: string }
        operator:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        createdAt: { type: string, format: date-time }
...